Fix missing assertion
