+++ pkg/testacc/4_helpers_test.go
@@ -1,0 +1,39 @@
+package testacc
+
+import (
+	"context"
+	"fmt"
+	"slices"
+
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/internal/provider"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+)
+
+func checkResourceOwnershipIsGranted(opts *sdk.ShowGrantOptions, grantOn sdk.ObjectType, roleName string, objectNames ...string) func(s *terraform.State) error {
+	return func(s *terraform.State) error {
+		client := TestAccProvider.Meta().(*provider.Context).Client
+		ctx := context.Background()
+
+		grants, err := client.Grants.Show(ctx, opts)
+		if err != nil {
+			return err
+		}
+
+		found := make([]string, 0)
+		for _, grant := range grants {
+			if grant.Privilege == "OWNERSHIP" &&
+				(grant.GrantedOn == grantOn || grant.GrantOn == grantOn) &&
+				grant.GranteeName.Name() == roleName &&
+				slices.Contains(objectNames, grant.Name.FullyQualifiedName()) {
+				found = append(found, grant.Name.FullyQualifiedName())
+			}
+		}
+
+		if len(found) != len(objectNames) {
+			return fmt.Errorf("unable to find ownership privilege on %s granted to %s, expected names: %v, found: %v", grantOn, roleName, objectNames, found)
+		}
+
+		return nil
+	}
+}
