+++ pkg/testacc/resource_grant_ownership_acceptance_test.go
@@ -3,19 +3,15 @@
 package testacc
 
 import (
-	"context"
 	"fmt"
 	"regexp"
-	"slices"
 	"testing"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config/providermodel"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/internal/provider"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 	"github.com/hashicorp/terraform-plugin-testing/config"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/plancheck"
-	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/hashicorp/terraform-plugin-testing/tfversion"
 )
 
@@ -1318,34 +1314,6 @@ func TestAcc_GrantOwnership_OnDatabaseRole(t *testing.T) {
 			},
 		},
 	})
-}
-
-func checkResourceOwnershipIsGranted(opts *sdk.ShowGrantOptions, grantOn sdk.ObjectType, roleName string, objectNames ...string) func(s *terraform.State) error {
-	return func(s *terraform.State) error {
-		client := TestAccProvider.Meta().(*provider.Context).Client
-		ctx := context.Background()
-
-		grants, err := client.Grants.Show(ctx, opts)
-		if err != nil {
-			return err
-		}
-
-		found := make([]string, 0)
-		for _, grant := range grants {
-			if grant.Privilege == "OWNERSHIP" &&
-				(grant.GrantedOn == grantOn || grant.GrantOn == grantOn) &&
-				grant.GranteeName.Name() == roleName &&
-				slices.Contains(objectNames, grant.Name.FullyQualifiedName()) {
-				found = append(found, grant.Name.FullyQualifiedName())
-			}
-		}
-
-		if len(found) != len(objectNames) {
-			return fmt.Errorf("unable to find ownership privilege on %s granted to %s, expected names: %v, found: %v", grantOn, roleName, objectNames, found)
-		}
-
-		return nil
-	}
 }
 
 func TestAcc_GrantOwnership_migrateFromV0941_ensureSmoothUpgradeWithNewResourceId(t *testing.T) {
